+++ only-danmuku-adapter/src/main/kotlin/edu/only4/danmuku/adapter/application/queries/statistics/GetTotalStatisticsInfoQryHandler.kt
@@ -1,7 +1,7 @@
 package edu.only4.danmuku.adapter.application.queries.statistics
 
 import com.only4.cap4k.ddd.core.application.query.Query
-import edu.only4.danmuku.application.queries._share.draft.statistics.PreviousDayStatistics
+import edu.only4.danmuku.application.queries._share.model.dto.Statistics.PreviousDayStatistics
 import edu.only4.danmuku.application.queries.statistics.GetTotalStatisticsInfoQry
 import edu.only4.danmuku.domain.aggregates.statistics.enums.StatisticsDataType
 import org.babyfish.jimmer.sql.kt.KSqlClient
@@ -26,7 +26,7 @@ class GetTotalStatisticsInfoQryHandler(
 
         // 按数据类型分组统计（合并）
         val countsByType = statisticsList
-            .groupingBy { StatisticsDataType.valueOf(it.dataType.toInt()) }
+            .groupingBy { StatisticsDataType.valueOf(it.dataType) }
             .fold(0) { acc, item -> acc + (item.statisticsCount ?: 0) }
 
         return GetTotalStatisticsInfoQry.Response(
