+++ only-danmuku-application/src/main/kotlin/edu/only4/danmuku/application/queries/category/GetCategoryListQry.kt
@@ -30,6 +30,6 @@ object GetCategoryListQry {
         /** 背景图 */
         val background: String? = null,
         /** 排序号 */
-        val sort: Byte
+        val sort: Int,
     )
 }
