+++ only-danmuku-application/src/main/kotlin/edu/only4/danmuku/application/queries/video_draft/GetVideoDraftCountByStatusQry.kt
@@ -22,6 +22,6 @@ object GetVideoDraftCountByStatusQry {
 
     data class Response(
         /** 数量 */
-        val count: Int
+        val count: Long,
     )
 }
